Issue24 delete image (#27)
* Download EXACT pre build

* Enabeld mrxs zip upload.

* Fixed right managment under linux

* MP4 and key bindings support (#23)

* Added key navigation for frames

* Bugfix GlobalAnnotations

* Fixed height bug

* Enumerate search and filter ID

* Fixed delete image bug

https://github.com/ChristianMarzahl/Exact/issues/24

* Sync search results between tabs

* Added basic registration

* Array as return values

+++ exact/exact/images/models.py
@@ -135,7 +135,7 @@ class Image(models.Model):
                     os.remove(str(path_temp))
                     self.filename = path.name
                 # Videos
-                elif Path(path).suffix.lower().endswith(".avi"):
+                elif Path(path).suffix.lower() in [".avi", ".mp4"]:
                     dtype_to_format = {
                                     'uint8': 'uchar',
                                     'int8': 'char',
@@ -258,7 +258,8 @@ class Image(models.Model):
                 self.objectivePower = 1
             self.save()
         except Exception as e:
-            os.remove(str(path))
+            if path.exists():
+                os.remove(str(path))
             raise
 
     def __str__(self):
